requirements.txt
fastapi==0.78.0
uvicorn==0.18.0
requests==2.26.0
transformers==4.21.1
scikit-learn==1.0.2
google-generativeai==0.1.0
torch==1.10.0
torchvision==0.11.1
numpy==1.21.0
pandas==1.3.0
pydantic==1.8.2
pytest==6.2.5
SQLAlchemy==1.4.23
